# Optimise concatenate along axis 0